Create quiz added in QuizCategories
